Added view for overall container and Docker information (#37) Added commands ``stop_unused``, ``remove_stopped`` and ``stop_all`` (#37) Moved container maintenance tasks from ``Container`` app to ``Kioscadmin`` app (#37)
