+++ containertemplates/templates/containertemplates/site_form.html
@@ -1,5 +1,5 @@
 {# Projectroles dependency #}
-{% extends 'projectroles/project_base.html' %}
+{% extends 'projectroles/base.html' %}
 {% load rules %}
 {% load projectroles_common_tags %}
 {% load crispy_forms_filters %}
