+++ containertemplates/templates/containertemplates/site_list.html
@@ -1,5 +1,5 @@
 {# Projectroles dependency #}
-{% extends 'projectroles/project_base.html' %}
+{% extends 'projectroles/base.html' %}
 {% load projectroles_common_tags %}
 {% load static %}
 {% load rules %}
@@ -96,4 +96,3 @@
 </div>
 
 {% endblock projectroles %}
-\
